+++ backend/team6/src/main/java/programmers/team6/domain/admin/dto/response/AdminCodeResponse.java
@@ -4,7 +4,5 @@ import java.util.List;
 
 import org.springframework.data.domain.Page;
 
-import programmers.team6.domain.member.dto.CodeReadResponse;
-
 public record AdminCodeResponse(Page<CodeReadResponse> codeReadResponse, List<String> groupCodes) {
 }
