+++ backend/team6/src/main/java/programmers/team6/domain/admin/dto/response/CodeDropdownResponse.java
@@ -1,4 +1,4 @@
-package programmers.team6.domain.member.dto;
+package programmers.team6.domain.admin.dto.response;
 
 public record CodeDropdownResponse(String code, String name) {
 }
