+++ backend/team6/src/main/java/programmers/team6/domain/admin/dto/response/CodeReadResponse.java
@@ -1,4 +1,4 @@
-package programmers.team6.domain.member.dto;
+package programmers.team6.domain.admin.dto.response;
 
 public record CodeReadResponse(Long id, String groupCode, String code, String name) {
 }
