+++ backend/team6/src/main/java/programmers/team6/domain/admin/repository/AdminVacationRequestSearchCustom.java
@@ -9,11 +9,11 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
 import programmers.team6.domain.admin.dto.response.AdminVacationSearchCondition;
 import programmers.team6.domain.admin.dto.response.VacationRequestSearchResponse;
+import programmers.team6.domain.admin.entity.Code_;
+import programmers.team6.domain.admin.entity.Dept_;
 import programmers.team6.global.querybuilder.CriteriaCustomPredicateBuilder;
 import programmers.team6.global.querybuilder.CriteriaCustomQueryBuilder;
 import programmers.team6.global.querybuilder.QueryUtils;
-import programmers.team6.domain.member.entity.Code_;
-import programmers.team6.domain.member.entity.Dept_;
 import programmers.team6.domain.member.entity.Member_;
 import programmers.team6.domain.vacation.entity.ApprovalStep;
 import programmers.team6.domain.vacation.entity.ApprovalStep_;
