+++ backend/team6/src/main/java/programmers/team6/domain/admin/repository/DeptRepository.java
@@ -1,4 +1,4 @@
-package programmers.team6.domain.member.repository;
+package programmers.team6.domain.admin.repository;
 
 import java.util.List;
 
@@ -7,13 +7,13 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import programmers.team6.domain.member.dto.DeptDropdownResponse;
-import programmers.team6.domain.member.entity.Dept;
+import programmers.team6.domain.admin.dto.response.DeptDropdownResponse;
+import programmers.team6.domain.admin.entity.Dept;
 
 public interface DeptRepository extends JpaRepository<Dept, Long> {
 
 	@Query("""
-		  SELECT new programmers.team6.domain.member.dto.DeptDropdownResponse(d.id,d.deptName)
+		  SELECT new programmers.team6.domain.admin.dto.response.DeptDropdownResponse(d.id,d.deptName)
 		  FROM Dept d
 		""")
 	List<DeptDropdownResponse> findAllDept();
