+++ backend/team6/src/main/java/programmers/team6/domain/admin/service/AdminService.java
@@ -12,8 +12,8 @@ import programmers.team6.domain.admin.dto.response.AdminVacationSearchCondition;
 import programmers.team6.domain.admin.dto.response.VacationRequestDetailReadResponse;
 import programmers.team6.domain.admin.dto.request.VacationRequestDetailUpdateRequest;
 import programmers.team6.domain.admin.repository.AdminVacationRequestSearchCustom;
-import programmers.team6.domain.member.entity.Code;
-import programmers.team6.domain.member.repository.CodeRepository;
+import programmers.team6.domain.admin.entity.Code;
+import programmers.team6.domain.admin.repository.CodeRepository;
 import programmers.team6.domain.vacation.entity.ApprovalStep;
 import programmers.team6.domain.vacation.entity.VacationRequest;
 import programmers.team6.domain.vacation.repository.ApprovalStepRepository;
