+++ backend/team6/src/main/java/programmers/team6/domain/admin/service/DeptService.java
@@ -1,13 +1,13 @@
-package programmers.team6.domain.member.service;
+package programmers.team6.domain.admin.service;
 
 import java.util.List;
 
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
-import programmers.team6.domain.member.dto.DeptDropdownResponse;
-import programmers.team6.domain.member.entity.Dept;
-import programmers.team6.domain.member.repository.DeptRepository;
+import programmers.team6.domain.admin.dto.response.DeptDropdownResponse;
+import programmers.team6.domain.admin.entity.Dept;
+import programmers.team6.domain.admin.repository.DeptRepository;
 import programmers.team6.global.exception.code.NotFoundErrorCode;
 import programmers.team6.global.exception.customException.NotFoundException;
