+++ backend/team6/src/main/java/programmers/team6/domain/member/repository/MemberSearchRepository.java
@@ -15,10 +15,10 @@ import jakarta.persistence.criteria.CriteriaQuery;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
 import lombok.RequiredArgsConstructor;
+import programmers.team6.domain.admin.entity.Dept_;
 import programmers.team6.global.querybuilder.CriteriaCustomPredicateBuilder;
 import programmers.team6.global.querybuilder.CriteriaCustomQueryBuilder;
 import programmers.team6.global.querybuilder.QueryUtils;
-import programmers.team6.domain.member.entity.Dept_;
 import programmers.team6.domain.member.entity.Member;
 import programmers.team6.domain.member.entity.Member_;
 import programmers.team6.domain.member.enums.Role;
