+++ backend/team6/src/main/java/programmers/team6/domain/member/util/mapper/MemberMapper.java
@@ -1,8 +1,8 @@
 package programmers.team6.domain.member.util.mapper;
 
 import programmers.team6.domain.auth.dto.request.MemberSignUpRequest;
-import programmers.team6.domain.member.entity.Code;
-import programmers.team6.domain.member.entity.Dept;
+import programmers.team6.domain.admin.entity.Code;
+import programmers.team6.domain.admin.entity.Dept;
 import programmers.team6.domain.member.entity.Member;
 import programmers.team6.domain.member.entity.MemberInfo;
 import programmers.team6.domain.member.enums.Role;
