+++ backend/team6/src/main/java/programmers/team6/domain/vacation/repository/VacationRequestSearchRepository.java
@@ -14,8 +14,8 @@ import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
 import lombok.RequiredArgsConstructor;
 import programmers.team6.global.querybuilder.CriteriaCustomQueryBuilder;
-import programmers.team6.domain.member.entity.Code;
-import programmers.team6.domain.member.entity.Dept;
+import programmers.team6.domain.admin.entity.Code;
+import programmers.team6.domain.admin.entity.Dept;
 import programmers.team6.domain.member.entity.Member;
 import programmers.team6.domain.vacation.dto.response.VacationRequestCalendarResponse;
 import programmers.team6.domain.vacation.entity.VacationRequest;
