+++ backend/team6/src/main/java/programmers/team6/domain/vacation/service/VacationService.java
@@ -17,10 +17,10 @@ import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import programmers.team6.domain.member.entity.Code;
-import programmers.team6.domain.member.entity.Dept;
+import programmers.team6.domain.admin.entity.Code;
+import programmers.team6.domain.admin.entity.Dept;
 import programmers.team6.domain.member.entity.Member;
-import programmers.team6.domain.member.repository.CodeRepository;
+import programmers.team6.domain.admin.repository.CodeRepository;
 import programmers.team6.domain.member.repository.MemberRepository;
 import programmers.team6.domain.vacation.support.MonthRange;
 import programmers.team6.domain.vacation.dto.request.VacationCreateRequestDto;
