+++ backend/team6/src/test/java/programmers/team6/domain/admin/repository/AdminVacationRequestSearchCustomTest.java
@@ -31,11 +31,9 @@ import lombok.extern.slf4j.Slf4j;
 import programmers.team6.domain.admin.dto.response.AdminVacationSearchCondition;
 import programmers.team6.domain.admin.dto.response.VacationRequestSearchResponse;
 import programmers.team6.domain.admin.enums.Quarter;
-import programmers.team6.domain.member.entity.Code;
-import programmers.team6.domain.member.entity.Dept;
+import programmers.team6.domain.admin.entity.Code;
+import programmers.team6.domain.admin.entity.Dept;
 import programmers.team6.domain.member.entity.Member;
-import programmers.team6.domain.member.repository.CodeRepository;
-import programmers.team6.domain.member.repository.DeptRepository;
 import programmers.team6.domain.member.repository.MemberRepository;
 import programmers.team6.domain.vacation.entity.VacationRequest;
 import programmers.team6.domain.vacation.enums.ApprovalStatus;
