+++ backend/team6/src/test/java/programmers/team6/domain/admin/repository/AdminVacationRequestSearchTestDataFactory.java
@@ -4,8 +4,8 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
-import programmers.team6.domain.member.entity.Code;
-import programmers.team6.domain.member.entity.Dept;
+import programmers.team6.domain.admin.entity.Code;
+import programmers.team6.domain.admin.entity.Dept;
 import programmers.team6.domain.member.entity.Member;
 import programmers.team6.domain.member.enums.Role;
 import programmers.team6.domain.vacation.entity.ApprovalStep;
