+++ backend/team6/src/test/java/programmers/team6/domain/admin/service/AdminServiceTest.java
@@ -23,8 +23,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import programmers.team6.domain.admin.dto.request.VacationRequestDetailUpdateRequest;
 import programmers.team6.domain.admin.repository.AdminVacationRequestSearchCustom;
 import programmers.team6.domain.admin.repository.AdminVacationRequestSearchTestDataFactory;
-import programmers.team6.domain.member.entity.Code;
-import programmers.team6.domain.member.repository.CodeRepository;
+import programmers.team6.domain.admin.entity.Code;
+import programmers.team6.domain.admin.repository.CodeRepository;
 import programmers.team6.domain.vacation.entity.ApprovalStep;
 import programmers.team6.domain.vacation.entity.VacationRequest;
 import programmers.team6.domain.vacation.enums.VacationRequestStatus;
