+++ backend/team6/src/test/java/programmers/team6/domain/auth/service/MemberRepositorySliceTest.java
@@ -12,13 +12,13 @@ import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabas
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
 import jakarta.transaction.Transactional;
-import programmers.team6.domain.member.entity.Code;
-import programmers.team6.domain.member.entity.Dept;
+import programmers.team6.domain.admin.entity.Code;
+import programmers.team6.domain.admin.entity.Dept;
 import programmers.team6.domain.member.entity.Member;
 import programmers.team6.domain.member.entity.MemberInfo;
 import programmers.team6.domain.member.enums.Role;
-import programmers.team6.domain.member.repository.CodeRepository;
-import programmers.team6.domain.member.repository.DeptRepository;
+import programmers.team6.domain.admin.repository.CodeRepository;
+import programmers.team6.domain.admin.repository.DeptRepository;
 import programmers.team6.domain.member.repository.MemberRepository;
 import programmers.team6.support.PositionMother;
