+++ backend/team6/src/test/java/programmers/team6/mock/MemberStub.java
@@ -2,8 +2,8 @@ package programmers.team6.mock;
 
 import java.time.LocalDateTime;
 
-import programmers.team6.domain.member.entity.Code;
-import programmers.team6.domain.member.entity.Dept;
+import programmers.team6.domain.admin.entity.Code;
+import programmers.team6.domain.admin.entity.Dept;
 import programmers.team6.domain.member.entity.Member;
 import programmers.team6.domain.member.enums.Role;
