+++ backend/team6/src/test/java/programmers/team6/support/MemberMother.java
@@ -2,7 +2,7 @@ package programmers.team6.support;
 
 import java.time.LocalDateTime;
 
-import programmers.team6.domain.member.entity.Dept;
+import programmers.team6.domain.admin.entity.Dept;
 import programmers.team6.domain.member.entity.Member;
 import programmers.team6.domain.member.entity.MemberInfo;
 import programmers.team6.domain.member.enums.Role;
